// translate-helper/src/utils/types.ts
enum types {
  ADD_LANGUAGE = 'ADD_LANGUAGE',
  REMOVE_LANGUAGE = 'REMOVE_LANGUAGE',
  CHANGE_LANGUAGE = 'CHANGE_LANGUAGE',
  SET_ACTIVE_FILES = 'SET_ACTIVE_FILES',
  CHANGE_ACTIVE_FILES = 'CHANGE_ACTIVE_FILES',
  SET_ACTIVE_LANGUAGE = 'SET_ACTIVE_LANGUAGE',
  CHANGE_LANGUAGE_ID = 'CHANGE_LANGUAGE_ID',
  CHECK_LANGUAGE_ID = 'CHECK_LANGUAGE_ID'
};

export default types;